transcriptor/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydub==0.25.1
vosk==0.3.45
pyannote.audio==3.1.1
python-multipart==0.0.6
numpy==1.24.3

whisper

onnxruntime
onnxruntime-gpu